Destructible objects now required to have an EventHandler for subscribing damage-related events. Character event handler now used by BattleUI to update player slots in Battle UI whenever character takes damage.

+++ Assets/Scripts/Battle/UI/HealthBarController.cs
@@ -14,7 +14,7 @@ public class HealthBarController : MonoBehaviour
 	
 	public void UpdateHealthBar (Character character)
 	{
-		float fill = character.HitPointsCurrent / character.HitPointsMax;
+		float fill = (float)character.HitPointsCurrent / (float)character.HitPointsMax;
 		Health.fillAmount = fill;
 		HealthVal.text = $"{character.HitPointsCurrent}/{character.HitPointsMax}";
 	}
